A lot of clean up

+++ src/collision.cpp
@@ -3,27 +3,27 @@
 #include "mesh.h"
 
 std::vector<physics::BoundingSphere> physics::allBoundingSpheres;
-std::vector<std::pair<size_t, size_t>> physics::allBoundingSpherenode;
-size_t physics::createBoundingSphere(BoundingSphere pBoundingSphere)
+std::vector<std::pair<unsigned int, unsigned int>> physics::allBoundingSpherenode;
+unsigned int physics::createBoundingSphere(BoundingSphere pBoundingSphere)
 {
 	allBoundingSpheres.push_back(pBoundingSphere);
 	return allBoundingSpheres.size() - 1;
 }
 
 
-size_t physics::linkNodeToBoundingSphere(size_t pNodeIndex, BoundingSphere pBoundingSphere)
+unsigned int physics::linkNodeToBoundingSphere(unsigned int pNodeIndex, BoundingSphere pBoundingSphere)
 {
-	size_t bs_index = allBoundingSpheres.size();
+	unsigned int bs_index = allBoundingSpheres.size();
 	allBoundingSpheres.push_back(pBoundingSphere);
 	return linkNodeToBoundingSphere(pNodeIndex, bs_index);
 }
 
-size_t physics::linkNodeToBoundingSphere(size_t pNodeIndex, size_t pBoundingSphereIndex)
+unsigned int physics::linkNodeToBoundingSphere(unsigned int pNodeIndex, unsigned int pBoundingSphereIndex)
 {
 	allBoundingSpherenode.emplace_back(pNodeIndex, pBoundingSphereIndex);
 	return allBoundingSpherenode.size() - 1;
 }
-float physics::checkTriangleIntersect(size_t pIndexOffset_A, size_t pEntity_A, size_t pIndexOffset_B, size_t pEntity_B)
+float physics::checkTriangleIntersect(unsigned int pIndexOffset_A, unsigned int pEntity_A, unsigned int pIndexOffset_B, unsigned int pEntity_B)
 {
 	using namespace mesh;
 	glm::uvec3 face_A;
@@ -37,7 +37,7 @@ float physics::checkTriangleIntersect(size_t pIndexOffset_A, size_t pEntity_A, s
 	glm::mat4 A_transform = entities::allMatrices[pEntity_A];
 	glm::mat4 B_transform = entities::allMatrices[pEntity_B];
 
-	for (size_t i = 0; i < 3; ++i) {
+	for (unsigned int i = 0; i < 3; ++i) {
 		face_A[i] = allIndices[pIndexOffset_A + i];
 		face_B[i] = allIndices[pIndexOffset_B + i];
 		
@@ -58,17 +58,17 @@ float physics::checkTriangleIntersect(size_t pIndexOffset_A, size_t pEntity_A, s
 	glm::vec3 distances[3] = { glm::vec3(), glm::vec3(), glm::vec3() };
 	
 	
-	for (size_t i = 0; i < 3; ++i) {
+	for (unsigned int i = 0; i < 3; ++i) {
 		distances[i] = n_B * A_verts[i] + (-n_B + B_verts[0]);
 		
 	}
 	return false;
 }
 
-float physics::checkSphereIntersect(size_t pSphereA, size_t pSphereB)
+float physics::checkSphereIntersect(unsigned int pSphereA, unsigned int pSphereB)
 {
-	std::pair<size_t, size_t> node_link_A = allBoundingSpherenode[pSphereA];
-	std::pair<size_t, size_t> node_link_B = allBoundingSpherenode[pSphereB];
+	std::pair<unsigned int, unsigned int> node_link_A = allBoundingSpherenode[pSphereA];
+	std::pair<unsigned int, unsigned int> node_link_B = allBoundingSpherenode[pSphereB];
 	BoundingSphere& A = allBoundingSpheres[node_link_A.second];
 	BoundingSphere& B = allBoundingSpheres[node_link_B.second];
