+++ src/collision.h
@@ -10,19 +10,19 @@ namespace physics {
 		glm::vec3 local_pos;
 		float radius;
 	};
-	extern std::vector<std::pair<size_t, size_t>> allBoundingSpherenode;//access this with a Bounding link index, first element of pair is an index to the node of a boundingsphere second is the bounding sphere index
+	extern std::vector<std::pair<unsigned int, unsigned int>> allBoundingSpherenode;//access this with a Bounding link index, first element of pair is an index to the node of a boundingsphere second is the bounding sphere index
 	extern std::vector<BoundingSphere> allBoundingSpheres;
-	size_t createBoundingSphere(BoundingSphere pBoundingSphere);
+	unsigned int createBoundingSphere(BoundingSphere pBoundingSphere);
 
-	size_t linkNodeToBoundingSphere(size_t pNodeIndex, BoundingSphere pBoundingSphere);
-	size_t linkNodeToBoundingSphere(size_t pNodeIndex, size_t pBoundingSphereIndex);
+	unsigned int linkNodeToBoundingSphere(unsigned int pNodeIndex, BoundingSphere pBoundingSphere);
+	unsigned int linkNodeToBoundingSphere(unsigned int pNodeIndex, unsigned int pBoundingSphereIndex);
 
-	float checkTriangleIntersect(size_t pIndexOffset_A, size_t pEntity_A, size_t pIndexOffset_B, size_t pEntity_B);
-	float checkSphereIntersect(size_t pSphereA, size_t pSphereB);
+	float checkTriangleIntersect(unsigned int pIndexOffset_A, unsigned int pEntity_A, unsigned int pIndexOffset_B, unsigned int pEntity_B);
+	float checkSphereIntersect(unsigned int pSphereA, unsigned int pSphereB);
 
 
 	template<class PrimA, class PrimB>
-	float checkIntersect(size_t pNode_A, size_t pNode_B, PrimA pA, PrimB pB)
+	float checkIntersect(unsigned int pNode_A, unsigned int pNode_B, PrimA pA, PrimB pB)
 	{
 		return 0.0f;
 	}
